feat(dtfs2-7430): split unit tests
